# internal/platform/apidocs/swagger.yaml
definitions:
...
  apierr.Envelope:
    properties:
      error:
        $ref: '#/definitions/apierr.EnvelopeEr'
      request_id:
        example: req_example
        type: string
      type:
        enum:
        - error
        example: error
        type: string
    required:
    - error
    - request_id
    - type
    type: object
  apierr.EnvelopeEr:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        example: input is required
        type: string
      type:
        enum:
        - invalid_request_error
        - authentication_error
        - permission_error
        - not_found_error
        - request_too_large
        - rate_limit_error
        - api_error
        - timeout_error
        - overloaded_error
        example: invalid_request_error
        type: string
    required:
    - message
    - type
    type: object
...
  memory.Conflict:
    properties:
      actual_version:
        type: integer
      content:
        type: string
      created_at:
        format: date-time
        type: string
      expected_version:
        type: integer
      id:
        type: string
      path:
        type: string
      store_id:
        type: string
    required:
    - actual_version
    - content
    - created_at
    - expected_version
    - id
    - path
    - store_id
    type: object
  memory.ConflictsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/memory.Conflict'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  memory.CreateStoreRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  memory.DeleteEntryRequest:
    properties:
      path:
        example: notes/project.md
        minLength: 1
        type: string
      version:
        description: 期望版本；不匹配返回 409。 || Expected version; a mismatch returns 409.
        type: integer
    required:
    - path
    type: object
  memory.EntriesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/memory.Entry'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  memory.Entry:
    properties:
      content:
        type: string
      deleted:
        type: boolean
      path:
        type: string
      sha256:
        type: string
      store_id:
        type: string
      updated_at:
        format: date-time
        type: string
      version:
        type: integer
    required:
    - content
    - deleted
    - path
    - sha256
    - store_id
    - updated_at
    - version
    type: object
  memory.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/memory.Store'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  memory.Revision:
    properties:
      content:
        type: string
      created_at:
        format: date-time
        type: string
      deleted:
        type: boolean
      id:
        type: string
      path:
        type: string
      session_id:
        type: string
      store_id:
        type: string
      version:
        type: integer
    required:
    - content
    - created_at
    - deleted
    - id
    - path
    - store_id
    - version
    type: object
  memory.RevisionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/memory.Revision'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  memory.Store:
    properties:
      created_at:
        format: date-time
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - created_at
    - id
    - name
    type: object
  memory.WriteEntryRequest:
    properties:
      content:
        type: string
      path:
        example: notes/project.md
        minLength: 1
        type: string
      version:
        description: 期望版本；新条目使用 0，更新使用当前版本，不匹配返回 409。 || Expected version; use 0 for
          new entries and the current version for updates. A mismatch returns 409.
        type: integer
    required:
    - path
    type: object
...
info:
  contact: {}
  description: |-
    Managed Agents 服务。所有 /v1 接口使用 Bearer Wave API Key 认证； || Managed Agents service. All /v1 endpoints require a Bearer Wave API Key;
    由 `wave bootstrap` 签发。Authorization 填写完整的 `Bearer <Wave API Key>`。 || issue it with `wave bootstrap`. Set Authorization to the full `Bearer <Wave API Key>` value.
    offset 分页返回 data 和可选 next_offset（省略表示末页）；消息历史使用 next_after，事件列表使用最后一条 sequence 续读。 || Offset pagination returns data and an optional next_offset (omitted on the last page); message history uses next_after, and event lists resume from the last sequence.
    JSON 错误返回 {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}。文件传输的 Range 错误返回 text/plain；条件下载和健康检查可返回空响应。 || JSON errors return {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}. File-transfer Range errors return text/plain; conditional downloads and health checks may return empty responses.
    JSON 请求拒绝未知字段，非 multipart 请求体上限为 2 MiB。可空字段通过 x-nullable 标记；省略与 null 不等价。 || JSON requests reject unknown fields; non-multipart request bodies are limited to 2 MiB. Nullable fields use x-nullable; omission is not equivalent to null.
    交互式文档见 /swagger/index.html，OpenAPI spec 见 /swagger/doc.json。 || Interactive documentation: /swagger/index.html. OpenAPI spec: /swagger/doc.json.
  title: Wave AI API
  version: "1.0"
paths:
...
  /v1/memory-stores:
    get:
      description: 分页返回当前所有者可见的记忆库。 || Return memory stores visible to the current
        owner, with pagination.
      operationId: memoryListStores
      parameters:
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.ListResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出记忆库 || List memory stores
      tags:
      - memory
    post:
      consumes:
      - application/json
      description: 为当前所有者创建一个命名的记忆库。 || Create a named memory store for the current
        owner.
      operationId: memoryCreateStore
      parameters:
      - description: 记忆库名称 || Memory store name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memory.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/memory.Store'
        "400":
          description: 请求体不是合法 JSON 或缺少必填字段 || Invalid JSON body or missing required
            fields
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 创建记忆库 || Create a memory store
      tags:
      - memory
  /v1/memory-stores/{id}/conflicts:
    get:
      description: 分页返回该记忆库因版本冲突被保留的冲突记录，包含期望版本、实际版本与冲突内容。 || Return retained version-conflict
        records for the memory store, with pagination, including expected and actual
        versions and conflicting content.
      operationId: memoryListConflicts
      parameters:
      - description: 记忆库 ID || Memory store ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.ConflictsResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 记忆库不存在或不在当前所有者范围内 || Memory store not found or outside the
            current owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出记忆冲突 || List memory conflicts
      tags:
      - memory
  /v1/memory-stores/{id}/entries:
    delete:
      consumes:
      - application/json
      description: 按乐观锁删除指定路径的记忆条目；版本不匹配时保留冲突内容并返回 409，产生一条删除修订记录。 || Delete a memory
        entry at the specified path using optimistic locking. A version mismatch retains
        the conflict and returns 409; a successful deletion creates a deletion revision.
      operationId: memoryDeleteEntry
      parameters:
      - description: 记忆库 ID || Memory store ID
        in: path
        name: id
        required: true
        type: string
      - description: 目标路径与期望版本 || Target path and expected version
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memory.DeleteEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Entry'
        "400":
          description: 请求体不是合法 JSON 或路径不是相对路径 || Invalid JSON body or non-relative
            path
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 记忆库不存在或不在当前所有者范围内 || Memory store not found or outside the
            current owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "409":
          description: 版本冲突，冲突内容已被保留 || Version conflict; conflicting content retained
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 删除记忆条目 || Delete a memory entry
      tags:
      - memory
    get:
      description: 按路径排序分页返回该记忆库的全部条目（含已删除标记的条目）。 || Return all entries in the memory
        store, including entries marked deleted, paginated and ordered by path.
      operationId: memoryListEntries
      parameters:
      - description: 记忆库 ID || Memory store ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.EntriesResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 记忆库不存在或不在当前所有者范围内 || Memory store not found or outside the
            current owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出记忆条目 || List memory entries
      tags:
      - memory
    put:
      consumes:
      - application/json
      description: 按乐观锁写入指定路径的记忆条目；版本匹配才生效并产生新版本与修订记录，版本不匹配时保留冲突内容并返回 409。 || Write
        a memory entry at the specified path using optimistic locking. A matching
        version creates a new version and revision; a mismatch retains the conflicting
        content and returns 409.
      operationId: memoryWriteEntry
      parameters:
      - description: 记忆库 ID || Memory store ID
        in: path
        name: id
        required: true
        type: string
      - description: 目标路径、内容与期望版本 || Target path, content and expected version
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memory.WriteEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Entry'
        "400":
          description: 请求体不是合法 JSON 或路径不是相对路径 || Invalid JSON body or non-relative
            path
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 记忆库不存在或不在当前所有者范围内 || Memory store not found or outside the
            current owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "409":
          description: 版本冲突，冲突内容已被保留 || Version conflict; conflicting content retained
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 写入记忆条目 || Write a memory entry
      tags:
      - memory
  /v1/memory-stores/{id}/revisions:
    get:
      description: 按修订 ID 倒序分页返回该记忆库的所有修订（含写入与删除）。 || Return all revisions of the
        memory store, including writes and deletions, paginated by revision ID descending.
      operationId: memoryListRevisions
      parameters:
      - description: 记忆库 ID || Memory store ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.RevisionsResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 记忆库不存在或不在当前所有者范围内 || Memory store not found or outside the
            current owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 记忆修订历史 || Memory revision history
      tags:
      - memory
...
securityDefinitions:
  BearerAuth:
    description: 填写完整的 Bearer <Wave API Key>，包含 Bearer 前缀和空格。 || Enter the full Bearer
      <Wave API Key> value, including the Bearer prefix and a space.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"